refactor: Move favicons discover to root, matching feeds/blogrolls
